+++ .config/config_example.yml
@@ -16,7 +16,3 @@ debug:
   enable: true
   profiling: false
   nossl: false
-
-extends:
-  workers: 4
-  process: 4
