+++ app/models/prisma/backend.prisma
@@ -1,0 +1,54 @@
+model BEConfig { // 略してベーコn(((
+    id                    String         @id @default("hol0")
+    host                  String         @unique
+
+    name                  String @default("Hol0")
+    description           String @default("An Interconnected Extensible Microblogging Platform🪐")
+
+    repositoryUrl         String  @default("https://github.com/hol0-dev/backend")
+    feedbackUrl           String  @default("https://github.com/hol0-dev/backend/issues")
+
+    admin                 String?
+    adminEmail            String? // 実はメールアドレス以外も入る。
+
+    maintainerName          String?
+    maintainerEmail         String?
+
+    impressumUrl          String?
+    tosUrl                String?
+    privacyPolicyUrl      String?
+
+    PushNotification      Boolean @default(false)
+    SWPublicKey           String?
+    SWPrivateKey          String?
+
+    enableMail            Boolean @default(false)
+    mailAddress           String?
+    smtpHost              String?
+    smtpPort              String?
+    smtpUser              String?
+    smtpPass              String?
+    smtpSSL               Boolean @default(false)
+
+    useObjectStorage      Boolean @default(false)
+    s3BaseUrl             String?
+    s3Bucket              String?
+    s3Prefix              String?
+    s3Endpoint            String?
+    s3Region              String @default("us-east-1")
+    s3AccessKey           String?
+    s3SecretKey           String?
+    s3useSSL              Boolean @default(false)
+    s3ForcePathStyle      Boolean @default(false)
+    s3setPublicRead       Boolean @default(false)
+
+    enableTurnstile       Boolean @default(false)
+    turnstileSiteKey      String?
+    turnstileSecretKey    String?
+
+    faviconUrl            String?
+    appleTouchIconUrl     String?
+    androidTouchIconUrl   String?
+
+    themeColor            String @default("#B0F7DD")
+}
