+++ app/models/prisma/channel.prisma
@@ -1,0 +1,26 @@
+model ChannelFollow {
+    followeeId String   @id
+    channelId  String
+
+    createdAt  DateTime @default(now())
+
+    channel    Channel  @relation(fields: [channelId], references: [id], onDelete: Cascade)
+    followee   User     @relation("channelFollowee", fields: [followeeId], references: [id], onDelete: Cascade)
+}
+
+model Channel {
+    id          String          @id
+
+    name        String
+    description String
+
+    remote      String?
+
+    createdAt   DateTime        @default(now()) @map("created_at")
+    updatedAt   DateTime        @default(now()) @updatedAt @map("updated_at")
+
+    host        Instance?       @relation(fields: [remote], references: [host])
+    
+    followers   ChannelFollow[]
+    notes       Note[]
+}
