+++ app/models/prisma/instance.prisma
@@ -1,0 +1,50 @@
+model Instance {
+    id                      String    @id
+    host                    String    @unique
+
+    usersCount              Int       @default(0)
+    notesCount              Int       @default(0)
+
+    name                    String?
+    description             String?
+
+    iconUrl                 String
+    faviconUrl              String
+    themeColor              String    @default("#ffffff")
+
+    firstRetrievedAt        DateTime?  @default(now())
+
+    isNotResponding         Boolean?  @default(false)
+    isSuspended             Boolean?  @default(false)
+    isBlocked               Boolean?  @default(false)
+    isSilenced              Boolean?  @default(false)
+
+    softwareName            String?
+    softwareVersion         String?
+    softwareHomepage        String?
+
+    openRegistrations       Boolean
+
+    adminName               String?
+    adminEmail              String?
+
+    maintainerName          String?
+    maintainerEmail         String?
+
+    infoUpdatedAt           DateTime  @default(now())
+    latestRequestReceivedAt DateTime  @default(now())
+
+    moderationNote          String?
+
+    langs                   String[]
+
+    tosUrl                  String?
+    privacyPolicyUrl        String?
+    inquiryUrl              String?
+    impressumUrl            String?
+    repositoryUrl           String?
+    feedbackUrl             String?
+
+    channels                Channel[]
+    user                    User[]
+}
