+++ app/models/prisma/schema.tmpl
@@ -1,0 +1,8 @@
+generator client {
+    provider = "prisma-client-py"
+}
+
+datasource db {
+    provider = "postgresql"
+    url      = env("DATABASE_URL")
+}
