+++ app/routes/__init__.py
@@ -1,7 +1,7 @@
-from robyn import SubRouter
+from fastapi import APIRouter
 
 from . import activitypub, avatar
 
-router = SubRouter("Holo")
+router = APIRouter()
 router.include_router(activitypub.router)
 router.include_router(avatar.router)
