+++ app/routes/activitypub/__init__.py
@@ -1,8 +1,8 @@
-from robyn import SubRouter
+from fastapi import APIRouter
 
 from . import well_known, nodeinfo, box, actors
 
-router = SubRouter("Holo")
+router = APIRouter()
 router.include_router(well_known.router)
 router.include_router(nodeinfo.router)
 router.include_router(box.router)
