--- src/process.py
@@ -1,3 +1,0 @@
-from app.server import app
-
-app.start(host="0.0.0.0", port=8080)
